+++ package.json
@@ -12,6 +12,7 @@
     "express": "^4.16.4",
     "helmet": "^3.15.0",
     "knex": "^0.16.3",
+    "morgan": "^1.9.1",
     "sqlite3": "^4.0.4"
   },
   "devDependencies": {
